separated context, removed auth, and small fixes
