+++ client/src/App.jsx
@@ -1,4 +1,4 @@
-import React, { Suspense, createContext, useState } from "react";
+import React, { Suspense } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import "@fontsource/poppins";
@@ -15,23 +15,18 @@ const HomePage = React.lazy(() => import("./pages/HomePage"));
 const SearchResultsPage = React.lazy(() => import("./pages/SearchResultsPage"));
 const AboutPage = React.lazy(() => import("./pages/AboutPage"));
 
-export const AuthContext = createContext([{}, () => {}]);
-
 const App = () => {
-  const [authUser, setAuthUser] = useState(null);
   return (
-    <AuthContext.Provider value={[authUser, setAuthUser]}>
-      <BrowserRouter>
-        <CssBaseline />
-        <ThemeProvider theme={theme}>
-          <Routes>
-            <Route path={"/"} element={withSuspense(HomePage)} />
-            <Route path={"/search"} element={withSuspense(SearchResultsPage)} />
-            <Route path={"/about"} element={withSuspense(AboutPage)} />
-          </Routes>
-        </ThemeProvider>
-      </BrowserRouter>
-    </AuthContext.Provider>
+    <BrowserRouter>
+      <CssBaseline />
+      <ThemeProvider theme={theme}>
+        <Routes>
+          <Route path={"/"} element={withSuspense(HomePage)} />
+          <Route path={"/search"} element={withSuspense(SearchResultsPage)} />
+          <Route path={"/about"} element={withSuspense(AboutPage)} />
+        </Routes>
+      </ThemeProvider>
+    </BrowserRouter>
   );
 };
