+++ client/src/contexts/FilterContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const FilterContext = createContext([{}, () => {}]);
+
+export default FilterContext;
