+++ client/src/pages/SearchResultsPage.jsx
@@ -1,9 +1,10 @@
-import React, { createContext, useState } from "react";
+import React, { useState } from "react";
 import { Stack, Typography } from "@mui/material";
 
 import BasePage from "./BasePage";
 import FiltersBar from "../components/FiltersBar";
 import SearchBar from "../components/SearchBar";
+import FilterContext from "../contexts/FilterContext";
 
 /**
  * https://github.com/BobaTalks/internship-tracker/issues/28
@@ -133,8 +134,6 @@ const MOCK_FILTER_DATA = {
     },
   },
 };
-
-export const FilterContext = createContext([{}, () => {}]);
 
 const SearchResultsPage = () => {
   const [filterData, setFilterData] = useState(MOCK_FILTER_DATA);
